+++ Modules/Microsoft365DSC/Modules/M365DSCDRGUtil.psm1
@@ -24,23 +24,6 @@ function Get-StringFirstCharacterToLower
     return $Value.Substring(0,1).ToLower() + $Value.Substring(1,$Value.Length-1)
 }
 
-function Remove-M365DSCCimInstanceTrailingCharacterFromExport
-{
-    [CmdletBinding()]
-    [OutputType([System.String])]
-    param(
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $DSCBlock
-    )
-
-    $DSCBlock = $DSCBlock.Replace("    ,`r`n" , "    `r`n" )
-    $DSCBlock = $DSCBlock.Replace("`r`n;`r`n" , "`r`n" )
-    $DSCBlock = $DSCBlock.Replace("`r`n,`r`n" , "`r`n" )
-
-    return $DSCBlock
-}
-
 function Rename-M365DSCCimInstanceParameter
 {
     [CmdletBinding()]